Restructuring the directories to reflect the overall project plan.
All tests passing locally
