+++ pyproject.toml
@@ -11,15 +11,17 @@ python = "^3.9"
 fastapi = "^0.105.0"
 uvicorn = "^0.25.0"
 python-multipart = "^0.0.6"
-python-dotenv = "^1.0.0"
 sqlalchemy = "^2.0.23"
-psycopg2-binary = "^2.9.9"
+
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.4.4"
 requests-mock = "^1.11.0"
 coverage = "^7.4.1"
 httpx = "^0.26.0"
+psycopg2-binary = "^2.9.9"
+python-dotenv = "^1.0.0"
+
 
 [tool.pytest.ini_options]
 filterwarnings = [
