+++ scripts/launch.sh
@@ -10,7 +10,7 @@ PORT=${1:-8000}
 echo "Launching app on port $PORT..."
 
 # Run the app via uvicorn on the specified port
-poetry run uvicorn api.api:app --reload --port $PORT &
+poetry run uvicorn file_transfer_api.src.api.api:app --reload --port $PORT &
 
 # Exit with the exit code of the uvicorn command
 trap "echo -e '\n App has been terminated...'; exit 0" SIGINT SIGTERM
